requirements.txt
crewai>=0.60
openai>=1.35
chromadb>=0.5.5
pypdf>=4
python-slugify>=8
PyYAML>=6
tiktoken
